one of last

+++ cut/models.py
@@ -4,8 +4,8 @@ from django.db import models
 class CutMusic(models.Model):
     cutting_music = models.FileField(upload_to='cut_music/')
     cutted_music= models.FileField(max_length=255,null=True,blank=True)
-    time_from = models.BigIntegerField(null=True,blank=True,default=0)
-    time_to = models.BigIntegerField(null=True,blank=True,default=30000)
+    time_from = models.FloatField(null=True,blank=True,default=0)
+    time_to = models.FloatField(null=True,blank=True,default=30000)
 class JoinMusic(models.Model):
     first_music = models.FileField(upload_to='documents/%Y/%m/%d')
     second_music = models.FileField(upload_to='documents/%Y/%m/%d')
